pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for runemu website for discovering races."
authors = ["sammy-may <samuel.james.may1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
rich = "^13.9.4"
logging = "^0.4.9.6"
django-extensions = "^3.2.3"
geopy = "^2.4.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
pytest-django = "^4.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
